pyproject.toml
[project]
name = "BoaBoard"
version = "0.1.0"
description = "BoaBoard is a modern, Python-powered attendance dashboard designed to streamline employee tracking and provide real-time insights with a sleek and intuitive interface."
authors = [
    {name = "Julius Koenig", email = "info@bastelquartier.de"},
]
dependencies = ["fastapi>=0.115.6", "mongoengine>=0.29.1", "uvicorn>=0.34.0", "typer>=0.15.1", "wiederverwendbar[full]>=0.6.3", "pydantic-settings>=2.7.1", "rich>=13.9.4", "IPython>=8.31.0", "openai>=1.59.7"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "GPL-3.0"}

[project.scripts]
boaboard = "boaboard.__main__:cli_app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "file"
path = "src/boaboard/__init__.py"
